# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Preloadet",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All User",
                "description": "Hamma Userlarni ko'rish",
                "operationId": "User_index",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "new User add",
                "description": "Yangi User qoshish",
                "operationId": "User_store",
                "requestBody": {
                    "description": "save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "middlename",
                                    "surname",
                                    "phone",
                                    "language",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "name"
                                    },
                                    "middlename": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "middlename"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "surname"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "+998901234567"
                                    },
                                    "language": {
                                        "type": "enum",
                                        "format": "text",
                                        "example": "uz"
                                    },
                                    "code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1254"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show User",
                "description": "bitta User hamma malumotlarini korsatadi",
                "operationId": "User_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "User_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Category",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "middlename",
                                    "surname",
                                    "phone",
                                    "language",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "name"
                                    },
                                    "middlename": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "middlename"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "surname"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "+998901234567"
                                    },
                                    "code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1254"
                                    },
                                    "language": {
                                        "type": "enum",
                                        "format": "text",
                                        "example": "uz"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "User_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/updete-lenguage/{lenguage}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "User_lenguage_update",
                "parameters": [
                    {
                        "name": "lenguage",
                        "in": "path",
                        "description": "lenguage",
                        "required": true,
                        "schema": {
                            "enum": [
                                "lenguage_uz",
                                "lenguage_ru",
                                "lenguage_eng"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "required": [
                    "name",
                    "middlename",
                    "surname",
                    "phone",
                    "language",
                    "code"
                ],
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "middlename": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "code": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}